backend/requirements.txt
fastapi==0.116.1
pydantic==2.10.6
pydantic_core==2.27.2
uvicorn==0.40.0
sqlalchemy
google-auth
google-auth-oauthlib
google-auth-httplib2
python-jose
python-dotenv
PyJWT
requests